+++ auramask/callbacks/__init__.py
@@ -1,0 +1,1 @@
+from auramask.callbacks.callbacks import init_callbacks  # noqa: F401
